Add multiscale pyramid generation methods to Volume class
